+++ fluent-react/examples/redux-async/src/reducer.js
@@ -1,0 +1,15 @@
+import { negotiateLanguages } from './l10n';
+
+export default function reducer(state = {
+  locales: negotiateLanguages(navigator.language)
+}, action) {
+  switch (action.type) {
+    case 'CHANGE_LOCALE':
+      return {
+        ...state,
+        locales: negotiateLanguages(action.value)
+      };
+    default:
+      return state;
+  }
+}