+++ fluent-react/examples/redux-sync/src/actions.js
@@ -1,0 +1,6 @@
+export function changeLocale(value) {
+  return {
+    type: 'CHANGE_LOCALE',
+    value
+  }
+}